Its reading JSON!!

+++ host-firmware/src/network/WiFiUtils.h
@@ -7,7 +7,7 @@ class WiFiUtilsClass{
 
 public:
     void printWiFiStatus();
-    int connectWiFI(char* ssid, char* password, int timeout,int attempts=1);
+    int connectWiFI(char* ssid, char* password, long timeout,int attempts=1);
 };
 
 extern WiFiUtilsClass WiFiUtils;
